+++ src/duplicates.c
@@ -135,7 +135,7 @@ duplicates_next(PyIUObject_Duplicates *self)
     PyObject *item=NULL, *temp=NULL;
     int ok;
 
-    while ( (item = (*Py_TYPE(self->iterator)->tp_iternext)(self->iterator)) ) {
+    while ( (item = Py_TYPE(self->iterator)->tp_iternext(self->iterator)) ) {
 
         /* Use the item if key is not given, otherwise apply the key. */
         if (self->key == NULL) {